New File adjustments
